Feat: 해시태그 추가/삭제 수정

+++ backend/src/main/java/com/ssafy/doit/service/GroupHashTagService.java
@@ -5,6 +5,7 @@ import com.ssafy.doit.model.GroupHashTag;
 import com.ssafy.doit.model.HashTag;
 import com.ssafy.doit.model.request.RequestGroup;
 import com.ssafy.doit.model.response.ResponseGroup;
+import com.ssafy.doit.model.user.User;
 import com.ssafy.doit.repository.*;
 
 import lombok.RequiredArgsConstructor;
@@ -20,7 +21,8 @@ import java.util.Optional;
 @Service
 @RequiredArgsConstructor
 public class GroupHashTagService {
-
+    @Autowired
+    private final UserRepository userRepository;
     @Autowired
     private final GroupRepository groupRepository;
     @Autowired
@@ -52,37 +54,51 @@ public class GroupHashTagService {
 
     // 그룹 정보 수정
     @Transactional
-    public void updateGroup(Group groupReq){
+    public int updateGroup(Long userPk, Group groupReq){
         Optional<Group> group = groupRepository.findById(groupReq.getGroupPk());
 
-        group.ifPresent(selectGroup ->{
-            selectGroup.setName(groupReq.getName());
-            selectGroup.setContent(groupReq.getContent());
-            selectGroup.setMaxNum(groupReq.getMaxNum());
-            selectGroup.setEndDate(groupReq.getEndDate());
-            groupRepository.save(selectGroup);
-        });
+        if(userPk == group.get().getLeader()){
+            group.ifPresent(selectGroup ->{
+                selectGroup.setName(groupReq.getName());
+                selectGroup.setContent(groupReq.getContent());
+                selectGroup.setMaxNum(groupReq.getMaxNum());
+                selectGroup.setEndDate(groupReq.getEndDate());
+                groupRepository.save(selectGroup);
+            });
+        }else
+            return 0; // 로그인한 유저가 그룹장이 아니면 수정불가
+        return 1;
     }
 
     // 그룹 해시태그 추가
     @Transactional
-    public int updateHashTag(Long groupPk, String hashtag){
-        Optional<HashTag> opt = hashTagRepository.findByName(hashtag);
-        if(opt.isPresent()) return 1;
-        else {
-            Group group = groupRepository.findById(groupPk).get();
-            findOrCreateHashTag(group, hashtag);
+    public int updateHashTag(Long userPk, Long groupPk, String hashtag){
+        Optional<Group> optGroup = groupRepository.findById(groupPk);
+        if(userPk == optGroup.get().getLeader()) {
+            HashTag hashTag =  hashTagRepository.findByName(hashtag).get();
+            Optional<GroupHashTag> gh = groupHashTagRepository.findByGroupAndHashTag(optGroup.get(),hashTag);
+            if(gh.isPresent()) return 2; // 이미 해시태그 존재
+            else {
+                Group group = groupRepository.findById(groupPk).get();
+                findOrCreateHashTag(group, hashtag);
+            }
         }
-        return 0;
+        else
+            return 0; // 로그인한 유저가 그룹장이 아니면 수정불가
+        return 1;
     }
 
     // 그룹 해시태그 삭제
     @Transactional
-    public void deleteHashTag(Long groupPk, String hashtag){
-        Group group = groupRepository.findById(groupPk).get();
-        HashTag hashTag =  hashTagRepository.findByName(hashtag).get();
-        GroupHashTag gh = groupHashTagRepository.findByGroupAndHashTag(group,hashTag).get();
-        groupHashTagRepository.delete(gh);
+    public int deleteHashTag(Long userPk, Long groupPk, String hashtag){
+        Optional<Group> optGroup = groupRepository.findById(groupPk);
+        if(userPk == optGroup.get().getLeader()) {
+            HashTag hashTag =  hashTagRepository.findByName(hashtag).get();
+            GroupHashTag gh = groupHashTagRepository.findByGroupAndHashTag(optGroup.get(),hashTag).get();
+            groupHashTagRepository.delete(gh);
+        }else
+            return 0; // 로그인한 유저가 그룹장이 아니면 수정불가
+        return 1;
     }
 
     // 해시태그 있으면 추가, 없으면 cnt +1
@@ -91,10 +107,10 @@ public class GroupHashTagService {
         // 입력한 해시태그들이
         Optional<HashTag> opt = hashTagRepository.findByName(hashtag);
         HashTag tag;
-        if(opt.isPresent()){ // 기존에 있으면 cnt +1
+        if(opt.isPresent()){        // 기존에 있으면 cnt +1
             tag = opt.get();
             tag.setCnt(tag.getCnt() + 1);
-        }else {                   // 기존에 없으면 새로 생성
+        }else {                     // 기존에 없으면 새로 생성
             tag = hashTagRepository.save(HashTag.builder()
                     .name(hashtag).cnt(1).build());
         }
